+++ src/views/pages/Login.vue
@@ -1,0 +1,12 @@
+<template>
+    <v-container
+      class="fill-height pa-0 ma-0"
+      style="background: #ffffff none repeat scroll 0% 0%;"
+      fluid
+    >
+      
+        <h1>LOGIN</h1>
+    </v-container>
+  </template>
+  
+